Django-B/pyproject.toml
[tool.poetry]
name = "django-b"
version = "0.1.0"
description = "django-project"
authors = ["zanwoll <skd03541@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.3"
djangorestframework = "^3.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
